api/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-ollama
chromadb
unstructured
pypdf
python-docx
gradio
requests
psycopg2-binary
markdown
beautifulsoup4